Add files via upload
2-What is a Centrifuge?
A centrifuge is a scientific device that is used to separate fluids, gasses, or liquids based on the density of the subject. The separation is acquired by spinning a container with the material at a very high speed. The resulting forces created will then cause the heavier materials to travel to the bottom of the container.

                                                  
3-How Centrifuges Work
A centrifuge works through the principle of sedimentation which is under the influence of gravitational force and centrifugal force and thus separating the substances based on their densities. The various types of separation techniques are known as isopycnic technique, pelleting technique, density gradient technique, phase separation and ultrafiltration techniques. Pelleting is the most common application of most centrifuges. At the base of the tube of the centrifuge, the particles are concentrated as a pellet and they are separated from the remaining solution and it is known as supernatant. Chemicals are converted during the separation phase from a matrix or aqueous solution into a solvent.









4-Types of Centrifuges ?
Benchtop centrifuges
Benchtop centrifuges are popular for their characteristic feature of requiring small space on the benchtop. Other features of benchtop centrifuges are the speed (RCF) of this centrifuge; it ranges from a few hundred to 50,000 x g and has interchangeable rotors; fixed angle, continuous flow, and swinging bucket rotors are available. Tubes of benchtop centrifuges range from less than 1 ml to a few liters.  
Microcentrifuges
As the name suggests, the microcentrifuges accommodate tubes containing small/micro volumes of samples like 2 ml, 1.2 ml, 0.5 ml, and tubes. So, these apply most frequently in the microbiological laboratory for separating nucleic acid and proteins. It typically spins at speeds up to 16,000x g, but in the case of specialized forms, the rate can reach up to 30,000 x g. Some models also have
Ultracentrifuges
The acceleration speed of ultracentrifuges is typically up to 1,00,000 x g, but the rate can reach 2,00,000 x g.
Vacuum centrifuges
Vacuum centrifuge uses centrifugal force, vacuum, temperature, and gas to remove liquid or gas from the sample to concentrate or desiccate the samples. Purifying nucleic acids, proteins, peptides, and other components used in research laboratories are some its uses.

Component
Box ( design )
Electric Motor 12 V
Circuit power supply 12V 
Sensor SMPS
Test Tube
Arduino
Timer